Move 2d and 3d pure python packages to submodules
