+++ src/cosserat_solver/__init__.py
@@ -1,6 +1,1 @@
 """Cosserat Solver Package."""
-
-# from .fourier import *  # Keep existing exports
-# from .greens_wrapper import get_greens_callback, FORTRAN_AVAILABLE
-
-# __all__ = ['get_greens_callback', 'FORTRAN_AVAILABLE']
