+++ src/cosserat_solver/dim2/_dispersion_core_2d_wrapper.py
