+++ src/cosserat_solver/dim2/_integrator_core_2d_wrapper.py
