+++ src/cosserat_solver/dim2/dispersion.py
