+++ src/cosserat_solver/dim3/__init__.py
@@ -1,0 +1,1 @@
+# 3d code
