+++ src/cosserat_solver/dim3/_cosserat_core_3d_wrapper.py
