+++ src/cosserat_solver/dim3/_dispersion_core_3d_wrapper.py
