+++ src/cosserat_solver/dim3/_elastic_core_3d_wrapper.py
