+++ src/cosserat_solver/dim3/dispersion_3d.py
