+++ src/cosserat_solver/dim3/elastic_3d.py
