+++ tests/unit-tests/test-comparisons/test_greens_wrapper.py
@@ -15,13 +15,13 @@ import pytest
 from mpmath import mp
 
 from cosserat_solver import consts
-from cosserat_solver._integrator_core_2d_wrapper import IntegratorFortran
+from cosserat_solver.dim2._integrator_core_2d_wrapper import IntegratorFortran
+from cosserat_solver.dim2.integrator import Integrator
 from cosserat_solver.greens_wrapper import (
     FORTRAN_AVAILABLE,
     evaluate_greens_fortran,
     evaluate_greens_python,
 )
-from cosserat_solver.integrator import Integrator
 
 # Test spatial positions
 POSITIONS = [
