+++ tests/unit-tests/test-python/test_cosserat_3d.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import numpy as np
 
-from cosserat_solver import cosserat_3d
+from cosserat_solver.dim3 import cosserat_3d
 
 
 def test_shape(material_parameters) -> None:
