+++ tests/unit-tests/test-python/test_dispersion_2d.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from mpmath import mp
 
 import cosserat_solver.consts as consts
-from cosserat_solver.dispersion import DispersionHelper
+from cosserat_solver.dim2.dispersion import DispersionHelper
 
 mp.dps = consts.TEST_PRECISION
